투표상세 normal / issue 분리

+++ backend/src/main/java/org/usyj/makgora/repository/NormalVoteOptionRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.usyj.makgora.entity.NormalVoteOptionEntity;
 
-public interface NormalVoteOptionRepository extends JpaRepository<NormalVoteOptionEntity, Long> {
-    List<NormalVoteOptionEntity> findByNormalVote_Id(Long normalVoteId);
+public interface NormalVoteOptionRepository extends JpaRepository<NormalVoteOptionEntity, Integer> {
+    List<NormalVoteOptionEntity> findByNormalVote_Id(Integer normalVoteId);
 }
