+++ backend/src/main/java/org/usyj/makgora/repository/NormalVoteRepository.java
@@ -3,5 +3,5 @@ package org.usyj.makgora.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.usyj.makgora.entity.NormalVoteEntity;
 
-public interface NormalVoteRepository extends JpaRepository<NormalVoteEntity, Long> {
+public interface NormalVoteRepository extends JpaRepository<NormalVoteEntity, Integer> {
 }
